[DS-707] Generify List usage
NB: There is a potential issue in the obtaining of list values - where the authority code was added, it changed a List of Strings to a List of String arrays ( the value, and the authority). Some of the code that used these methods did not take into account that there was a String[] and would have thrown a classcastexception.

This has now been accounted for, but in doing so, it ignores the authority value component of the array. We need to consider if this needs to be updated, or possibly if that method should just be removed (doesn't appear to be used)

git-svn-id: http://scm.dspace.org/svn/repo/dspace/trunk@5601 9c30dcfa-912a-0410-8fc2-9e0234be79fd