DS-712 : DSpace Fails to Check if a Handle is already assigned before assigning a new handle
The resolution to this problem is not to for DSpace to query all handles when assigning a new one (which can be taxing for large installs).  Rather this fix changes the 'update-sequences.sql' scripts (for both Postgres & Oracle) such that the Handle sequence number will always be updated to ensure DSpace will assign a unique handle on the next handle request.  I've also documented in the AIP backup/restore docs (DS-466) that update-sequence.sql script should always be run after a full AIP restore (which is the primary scenario where DSpace may encounter handle assignment issues).

git-svn-id: http://scm.dspace.org/svn/repo/dspace/trunk@5691 9c30dcfa-912a-0410-8fc2-9e0234be79fd