Clear index flag only possible under the import action.
After some internal discussion, we establish that it is correct to avoid
clearing index at clean-cache operation. So, by this restriction one
also state that is only possible to re-import all items and not just
clean all items (making OAI interface incoherent with the past).

As Ivan Masár stated at #37, it must be possible to throw an 503 HTTP
status code when re-importing all items to provide an even correct OAI
interface.