DS-1483 Generate citation_pdf_url in more cases
Google Scholar uses the citation_pdf_url header tag on item pages to determine
which of an item's bitstreams to link to (if any). Standard DSpace behaviour is
to include this link only if the item has exactly one bitstream in the ORIGINAL
bundle and if this item is of type application/pdf. The desired behaviour for
the LCoNZ IRRs is this:
 - if the item has only one bitstream in the ORIGINAL bundle, link to this one
 - if the item has a primary bitstream in the ORIGINAL bundle, link to this one
   (regardless of how many bitstreams there are in total in the ORIGINAL bundle).
Additionally, the bitstream link should only appear when the bitstream selected
according to the rules above is publicly visible.
The mimetype of the bitstream is not taken into account, as per Google Scholar
feedback that this is no longer as important to them.