Remove broken reporting functionality
This file computes reports for data packages. Part of those stats
depend on accessing the data files that are part of the package. Since
the API for looking up files based on their identifiers has changed,
this portion of the code is currently broken.

I'm commenting it out rather than fixing it, because (a) we don't need
this functionality right now, and (b) the fix would take longer than I
have available for running a needed report. The next time such a
report is needed, we will need to reinstate the portion that looks up
data files.