Provide normalized DOIs in hasPart and isPartOf relationships.
This is a bit of a hack. Since Dryad still stores handles in the
isPartOf relationships, we make use of the structured DOIs to
construct the correct isPartOf DOI. The assumption here is that we
will eventually fix the isPartOf metadata field and then update this
code to use it.