Re-instate package statistics for embargoes and sizes
The old code for package statistics was a quick hack that used OAI and
METS to calculate package-level embargo and size information. Removed
this old code and replaced it with proper calls to the DSpace data
model.

Also improved debugging output to indicate why items are being skipped
when the package statistics are generated.