Additions to workflow to support registration of blacked out items
Added a step registerPendingPublicationStep before pendingPublicationStep and an action RegisterPendingPublicationAction that registers the DOI.
This way, the path through the workflow determines how to register the DOI.
Instead of checking database for a specific state, now adding provenance metadata to indicate blackout was entered.
This provenance metadata is used to determine the target and metadata to register the DOI with.