DS-1561 build.properties breaks no-auth SMTP
A limitation in build.properties filtering requires us to specify empty values of properties instead of commenting them out, as documented below:

https://wiki.duraspace.org/display/DSDOC3x/Configuration#Configuration-Thebuild.propertiesConfigurationPropertiesFile

This caused a problem with the mail.server.username and mail.server.password properties, which behaved differently when they were empty and when they were not present (commented out).

This commit fixes them by making the case of empty mail.server.username behave the same as if it was missing.