Move View File Details link
View File Details link was previously on the line with the download
link, which was based on METS metadata for the file.  Embargoed items
hide this metadtata, so the field wasn't rendered.

Moved "File Details" to a new table row, since it doesn't depend on the
mets metadata, just the file's DOI.