Update namespace for bibo ontology
When we need to combine our metadata with metadata from external
entities, we want to avoid te possiblity of conflicting XML schemas
for a given namespace. Since Dryad does not want to become the holder
of the "official" bibo schema, we're declaring our own bibo namespace
to reference our schema for the subset of bibo we use.