Don't add dc.date.accessioned when entering blackout
This resulted in two dc.date.accessioned values for items that went through publication blackout.
It was only added to provide a publication date year for items when registering blackout metadata.
Now, blackout items are registered with a publication date of 9999 (since (:tba) fails the XSD validation)
So it is not necessary to add a dc.date.accessioned before the item is really approved