Performance and UI improvements for itemmap
This commit remove hardcoded reference to the author index for both discovery
and lucene provider. The admin is now able to perform a generic search or choose
from one of the customized indices or filters.
A very simple pagination, only next and previous buttons, has been introduced to
avoid performance issues with large result set.
Finally has been introduced optimization to perform the "Not owning collection"
check and, limitly to discovery, authorization check directly as part of the
search.