DataCiteConnector: remove cache.
We introduced the cache to get better performance while reserving and
registering DOIs. To be able to process new submissions independently
from the status of the registration agency API we introduced a cron job
that registers and reserves doi. It is not important if the cronjob is
some seconds faster or not. The cache brought some new problems and
makes the code more complicated. As it is not needed any more I removed
it.