Move addPropagateButton template to forms.xsl to prevent broken dependency
Originally this template was in DryadUtils.xsl, which is imported by Mirage.
However, the Paypal.xsl theme does not import DryadUtils.xsl, and forms.xsl was referencing a template that didn't exist.

The template has been moved to the same file that calls it, so that other themes don't depend on DryadUtils.