Correct total number of log records when count=0
When the DataONE log aggregator is requesting log entries, it starts
by making a request like /mn/log?count=0. The aggregator expects the
response document to contain zero rows, but still contain a correct
total for the number of rows that matched the query. This expectation
aligns with standard SOLR behavior.

Our logging code was erroneously short-circuiting this behavior and
returning a result document that always declared total=0. This commit
removes the short-circuit.