do not delete checked authors/keywords unless "remove selected" button clicked
The org.dspace.submit.step.getRepeatedParameter() method previously filtered out
any checked items for an list that permits removal (for example, an author or
keyword list on the page used to describe publication by inputting metadata),
regardless of the user's intention to add a new item or remove existing items
from that list.

This commit updates the filter by only filtering out checked items if the button
clicked to initiate the Submit request is a "Remove" button for the given
metadata item, per the button's input/name value.