removed source-map path from jquery-migrate-min
The jquery-migrate-1.1.1.min.js file had a source-map file reference
that was causing the Safari debugger not to be able to step through
code. Since the jquery-migrate plugin is not likely to be debugged,
I removed the file reference to the source map file. I could not
find a version of the min.map file online at the correct version
number.