Moved review token utils to new class
Originally these static methods were added to SolrLogger, but testing
them causes SolrLogger to load, which depends on ConfigurationManager.
While there are strategies to load ConfigurationManager in testing, it
wasn’t in line with the purpose of this changeset, so I have instead
moved these utilities to a SolrLoggerUtils class, which has no
configuration dependencies and is therefore more testable.
. Since the new functionality here is actually