factor describe-publication form element div.ds-previous-values
The outputting of an element div.ds-previous-values in the
stylesheet webapp/xmlui/themes/Mirage/lib/xsl/core/forms.xsl
has been replaced with a call to a single named template to
handle the element. This factoring is meant to help in simplifying
the later update of the user-input fields.

The new named template has one parameter of type xalan:node-set,
so the namespace handling for the stylesheet was also updated.