finalize author-edit functionality for first pull request
This commit finalizes the js, xsl, css, and java commits required for
implementing the trello-71 functionality for a pull-request review.

The bulk of this commit is in the Mirage theme's dryad-pages.js file,
which has been extented to handle re-ordering and editing of author
names. The JS changes in this commit involve event handlers for the
Edit and Remove buttons on the describe-publication page.

The Miarage theme forms.xsl file was also modified to support the
Edit/Remove updates.

This commit was hand-tested against:

    Chrome 35.0
    Firefox 30.0
    Safari 7.0.5
    Opera 22.0