Batch import into multiple collections
If no -c argument is given ItemImport checks if a file named
'collections' exists inside the directory of each item.

This file must contain either the handles or the numeric ids of the
collections, one per line. The item is added to all those
collections. The collection specified in the first line is the owning
collection.