Uses all possible identifiers to generate URIs.
This commit enables dspace-rdf to use all PersistentIdentifiers minted
by DSpace and not only handles. It makes advantage of the changes
introduced to DSpace by DS-1990. Persistent Identifers should and will
be used as fully functional http URIs only (e.g. a DOI will be used in
the form http://dx.doi.org/<doi> and not as doi:<doi>).