LIBDRUM-324. add new <dcterms:record> root element for QDC metadata
The OAI-PMH [1] specification calls for a single root tag under the metadata element, which is why
in our problem examples the first tag <dc:title> is accepted and then the second <dc:creator> is rejected.

The QDC XSD [2]  allows for a tag with any name to be used as the root tag:  "This is included as
a convenience for schema authors who need to define a root or container element for all of the DC
elements and element refinements".  I chose <dcterms:record> as the container element.

The mapping now passes validation [3].

[1] http://www.openarchives.org/OAI/openarchivesprotocol.html#Record
[2] http://dublincore.org/schemas/xmls/qdc/dcterms.xsd
[3] http://validator.oaipmh.com/

https://issues.umd.edu/browse/LIBDRUM-324