LIBDRUM-366. add "DRUM Customizations" side heading to sidebar
FORMATTING COMMIT
This commit is due to formatting differences between the Dspace formatter for eclipse and the formatting of the source files.
The file is the same, the only difference is the formatting. This was done so that the diffs for the actual changes are visible clearly in later commits.

https://issues.umd.edu/browse/LIBDRUM-366