Change symlinks into actual files
It was useful to create a test directory using symlinks, so the files
only needed to be in one place at a time. However, the symlinks fail
to work when the source directory and installation directory have
different structures. It is more straightforwrd to maintain copies of
these files in the test config.