Make proper 'Using DryadLab Modules' page from DOC
This is a simple translation of our shared Word doc
(https://drive.google.com/a/ibang.com/file/d/0B6vgl7VhMIDiQi1UUWZ2MEg3aEE/view)
to a static web page in the DryadLab theme. I've added a link in the main site
nav, but it can of course also be linked from elsewhere in the site.