Guard against almost-empty strings
In some circumstances, the metadata supplied on the initial
publication page will consist of spaces. This occurred recently for
the prism_publicationName_authority due to changes in the authority
control system. This commit trims the strings that are entered on this
page, and verifies that the strings have content before using them.