DS-2731 Don't send event to GA if DNT:1 header is present
The Do Not Track header (https://en.wikipedia.org/wiki/Do_Not_Track) is
designed to let people opt-out of third-party tracking (such as GA) and
Firefox at least now has an easy way to set this in the preferences. The
Google Analytics event listener should honour the Do Not Track header
and not send the event to GA when the header is present and set to 1.