Set default search operator to AND
Setting Solr's default search operator to AND for the Discovery index leads to
search behaviour that's closer to what our users expect -- "do what Google
does". There is a configuration option for the search operator in dspace.cfg.
However, that setting is used only by the legacy DSQuery class for the now
deprecated Lucene-based search.