DS-2892 Harvested Items weren't actually persisted
During collection harvest, the retrieved items were never committed to the db.
The context needed to be completed for the items to persist.
During the reimport (which first deletes the items), the harvested items were still being referenced, resulting in a foreign key constraint violation.
First, check and delete the harvested_item if necessary