[DS-1494] Permit a site to leave administrative credentials unconfigured.
This enables the use of a separate 'jndi.properties' file, which can
be maintained separately from DSpace, kept out of source repositories
and patch submissions, etc.

More debugging.

Begin disentangling user search (which requires administrative login),
user credential testing (which requires login as the user), and
fetching of user attributes (which is currently bound up with user
search and does not happen if the user is not searched for).