Make the code for Submit button one line
Removing the whitespace within the XSL that generates the "Submit Now"
button. This should produce no visible changes in a running
environment. The whitespace needs to be removed so the secondary
server can use Apache's mod_substitute to match and replace this text
when disabling the button.

(This is not a great long-term solution for creating a read-only
server, but it's the solution we've been using. A more thorough
solution is described on https://trello.com/c/Qt7kCaxw.)