[DS-3482] Accept any type of DBConnection.
DSpace.getSingletonService(Class) looks for a service of type Class
named Class.getName().  I don't see how this ever worked, since
DBConnection is an interface, and an implementation will have some
other name.  Instead request any service which is instanceof
DBConnection.  There should be only one.

This patch uncovers a slew of new NPEs in dspace-api tests, which are
happening in dspace-services.  We have multiple Ehcache CacheManager
instances for some reason, and newer versions of Ehcache throw a fit
when you try to create multiple managers with the same configuration.