vocabulary-support: Fix reusing of the same dialog
If the same vocabulary is used multiple times on the same input page, the
dialog is only fetched the first time from the server and reused afterwards.
The problem is, that the target input field is contained in the dialog. When
reusing the dialog the target field should be updated.