AuthorityIndexer: Remove state
The DSpaceAuthorityIndexer now returns all AuthorityValues for the specified
Items at once. It iterates over all values of all metadata fields (that are
configured for authority) of the item and gets or generates the authority
values for all of them. So it can be called from various threads without the
possibility for concurrent data access.

This currently removes the cache for newly generated authority values.