BitstreamStorageServiceImpl: Do not call update
If the BitstreamStorageService calls update on the just cloned bitstream, this
bitstream will not belong to any Item and the user may get authorization
failures.

The update call can be removed here, because the clone method is only used by
the AbstractVersionProvider and it will call update on the bitstream later by
itself.