DS-4169: Treat empty language as null in edit item
The metadata editor in the admin edit item view always submits an empty string
if there is no language (the HTML input fields cannot submit null values). To
preserve the null values of the language attributes all empty strings in the
language input fields are treated as null values, now.