DS-4506 Add a limit on the maximum value allowed for the parameter rpp (resuts per page) in browse and search pages.
If this value is not controlled, all the items can be returned in a single query, which could cause a DOS if an intentional attack is performed.
In solr, a variable MAX_RESULTS_ALLOWED to control the max rpp value was created, while in browse page there is a check for the entered value is not greater than the maximum value shown in the select used to set the parameter.