Better fix for configuration issue in c463b632dd24e0c017d88b36b2b6fa0f13d95fb1
Other code now depends on the 'id' of the curation script configuration bean,
so it can't be renamed.  Instead, remove the bean post processing code that
pokes 'dspace.cfg' properties into a bean which are prefixed with the bean's
'id', since nobody can remember why we do that and it causes subtle magical
problems.  It may be for Guice, which is no longer an option for DSpace.