Speed up Docker image rebuilds with BuildKit and better layer reuse
Using BuildKit, we can parallelize the build of the multi-stage
Dockerfile by moving the things that do not require the DSpace
source code to the beginning of each stage. This includes the
installation of the basic build tools, such as Node, npm, or Grunt.

This commit speeds up the rebuild of Docker images by reordering
the various stages, using layers to preinstall the frontend tools,
and using a new stage to cache the Maven dependencies in its own
layer. Due to limitations in the Dockerfile format, this requires
listing each and every POM file in the project, but hopefully
this shouldn't require changing very often.

The commit also adds a convenience compose.sh script for using
Docker compose, so you can simply do this:

    ./compose.sh build
    ./compose.sh up

The commit needs to explicitly set up the GEM_HOME and GEM_PATH
environment variables, because the Grunt execution is given those
variables and the build fails with an empty GEM_PATH. Alternatively,
we could drop the mention of GEM_HOME and GEM_PATH in the
dspace/modules/xmlui-mirage2/pom.xml file.