dspace-xmlui-mirage2: use a static copy of modernizer.min.js
Long story short: this was previous installed by bower, but bower
is deprecated and we are trying to move dependencies to npm. The
version here is 2.8.3, which is six years old and does not exist
on npm's registry, and can't be installed from git because it is
missing package.json. Since the file is not changing whatsoever
I think it's fine if we just keep a static copy in the tree and
bundle it in with the theme.js using grunt, which has the added
benefit of doing one less network request.