LIBDRUM-650. Excluded all Java files in "additions" module
Excluding just the Java files with compile errors was not sufficient
to enable the Maven build, because some of the classes without compile
errors depend on the UMD-specific classes that had compile errors.

Excluding all Java files in the "additions" module, so that files can
be added back one-by-one as needed.

https://issues.umd.edu/browse/LIBDRUM-650