LIBDRUM-665. Ensure that stock discovery.xml config is used for testing
Added a stock DSpace 7.2.1 copy of the
"dspace/config/spring/api/discovery.xml" file to the
"test/data/dspaceFolder" of the "dspace-server-webapp" application so
that it would be used in place of the UMD-customized version when
running the tests.

It is necessary to place the "stock" file here, because there appears
to be no way to modify the Discovery functionality without
customizing the actual "config/spring/api/discovery.xml" file.

https://issues.umd.edu/browse/LIBDRUM-665