LIBDRUM-650. Remove private/obsolete dependencies
It is not clear if GitHub Actions can interact with a private Maven
repository so removed the UMD-specific "lib-util" package, and updated
the "marc4j" package to a version available from Maven Central.

The only class the code uses from the "lib-util" package is the
"ErrorHandling" class, which seems to be a really old class intended to
provide some functionality that log4j was presumably missing at the
time. It can likely be replaced with existing log4j functionality.

The "marc4j" package is available from Maven Central, but v2.0 is no
longer available. Setting to the latest version is v2.9.2, which
(hopefully) has no breaking changes.

https://issues.umd.edu/browse/LIBDRUM-650