LIBDRUM-672. Stub/nonfunctional implementation of "Unit" entity
Created a "stub" (and nonfunctional) implementation of
the "Unit" database entity, to have a unit test for verifying
that testing it working in the "additions" module.

Updated XML Spring configuration files in root "dspace/config"
to include the "Unit" entity.

Placed "stock" versions of XML Spring configuration files into the "dspace-api" module, to ensure that that module would use the stock files, instead of the UMD-modified ones.

https://issues.umd.edu/browse/LIBDRUM-672