LIBDRUM-674. Added "start-handle-server-fg" Handle server script
The "k8s-drum" configuration is expecting to start the built-in Handle
server using a "dspace/bin/start-handle-server-fg" script.

This script is a replacement for the "start-handle-server" script, and
is needed for the Kubernetes configuration, so that the "handle-server"
pod will run the Handle server as a foreground job, instead of as a
background daemon process (which is what the "start-handle-server"
script does).

The "/dspace/bin/start-handle-server-fg" script exists in the
"docker.lib.umd.edu/drum:6.3-drum-7.1" Docker
image, but the script is not present in the Git repository, so it is
unclear when/how it was added, although it possibly came from the
"MDSOAR" configuration at some point.

https://issues.umd.edu/browse/LIBDRUM-674