LIBDRUM-657. Refactored CAS/LDAP code to simplify testing
Created an "Ldap" interface, and moved existing code in the "Ldap"
class into an "LdapImpl" class.

In "CASAuthentication", created an "authenticate" method that takes
an Ldap object directly, instead of creating it within the method.
The unit tests for "authenticate" run against this version of the
method, and the original "authenticate" method delegates to it.

Added "spring-test" dependency to additions/pom.xml to get access to the
"MockHttpSession" class.

https://issues.umd.edu/browse/LIBDRUM-657