LIBDRUM-657. Improved EPerson registration
Improved EPerson registration to better follow example in
ShibAuthentication.

In LdapImpl, turned off authorization system (as in done in
ShibAuthentication) to create a new user, instead of using a known
"admin" account.

This also corrected an issue where a newly created user was not
immediately logged in to the front-end on the initial login
after creation.

https://issues.umd.edu/browse/LIBDRUM-657