LIBDRUM-657. Ensure "service" URL uses "http" protocol
When testing in Kubernetes, CAS authentication was failing because
of the "service" URL initially provided to CAS was using the "https"
protocol, and when validating the CAS ticket, the "service" URL
derived from the incoming request (in the "getServiceUrlFromRequest"
method) used the "http" protocol, leading to an "service mismatch"
error.

The incoming request uses the "http" protocol presumably because
Kubernetes is using Nginx to handle HTTPS and then passing the request
to the application as an HTTP request. So added code to ensure that
the initial CAS request uses "http" for the service URL.

Note: In DSpace 6, the "service" URL in the initial CAS request was
also "http", although in that case it was directly calculated from the
login endpoint (with a substitution of "cas-login" for "login), so
this code was not necessary, as the "service" URL was already "http".

https://issues.umd.edu/browse/LIBDRUM-657