LIBDRUM-692. Remove old DSpace files no longer in stock DSpace 7.3
Compared the files in the stock DSpace "dspace-7.3" repository to the
files in the "umd-lib" repository using the "diff" command:

```bash
$ diff --brief --recursive  ~/dspace-7.3/DSpace/ drum
```

and examined each of the files "Only in drum" to see if they could be
removed. Files were removed, using the criteria:

* They had previously existed in a DSpace version, but were no longer
  present in DSpace 7.3

* The files did not appear to contain any UMD customizations

If there was any doubt about the continued validity of a file, the file
was kept.

https://issues.umd.edu/browse/LIBDRUM-692