LIBDRUM-692. Updated UMD custom code to use log4j v2 classes
Modified UMD custom code to use log4j v2 classes, instead of log4j v1
classes. This is in line with most of the rest of the DSpace code, and
is necessary because "log4j-1.2-api" bridge library is no longer
part of the upstream POM configuration as a "compile" dependency.

https://issues.umd.edu/browse/LIBDRUM-692