LIBDRUM-695. Ensure CNRI Handle server has necessary dependencies
In the "dspace-api/pom.xml" file, the Jetty dependencies for the
"cnri-servlet-container" (the CNRI Handle Server) are excluded because
they conflict with Jetty versions provided by DSpace.

In DSpace 7.3, the Jetty dependencies used by the
"cnri-servlet-container" are still excluded, but no other DSpace
dependency includes the Jetty dependencies as "compile" dependencies
(they are included as "test" dependencies, which does not get them
included in the the final assembly package).

To ensure that the CNRI Handle Server has the necessary dependencies,
adding them as explicit dependencies (using the DSpace Jetty version)
to the "dspace/modules/additions/pom.xml" file.

The need for these dependencies should be re-evaluated when new DSpace
releases are made.

https://issues.umd.edu/browse/LIBDRUM-695