LIBDRUM-684. Added optional upload step for for MHHEA
In "access-conditions.xml" added a custom UploadConfiguration bean that
allows uploads to be optional, then added the custom configuration as
an entry in the uploadConfigurationService bean "map".

Note: the "key" value in the "entry" tag, must match the "id" of the
"step-definition" field in "item-submission.xml".

In "item-submission.xml", added "optionalUpload" step-definition,
and added the step to the "MHHEA_Collection" submission process.

https://issues.umd.edu/browse/LIBDRUM-684