LIBDRUM-705. Removed Jetty dependencies from additions/pom.xml
In the stock DSpace 7.3, various Jetty dependencies needed for the
CNRI Handle server were excluded in the "dspace-api/pom/xml".

In DSpace 7.4, these Jetty dependencies appear to have been re-included
so the inclusion of the Jetty dependencies as part of the
"dspace/modules/additons/pom.xml" file are no longer necessary.

https://issues.umd.edu/browse/LIBDRUM-705