LIBDRUM-710. Added log message describing duplicate title
Modifed "EtdLoader" to display a warning login message when a
duplicate title is found (in addition to sending an email). This is
to enable the possibility of finding items with duplicate titles in
the log, instead of relying solely on email messages.

Also modified an exception handler to provide a stack trace when
receiving an invalid file.

https://issues.umd.edu/browse/LIBDRUM-710