LIBDRUM-716. Restore UMD-customized XmlWorkflowItem class
In LIBDRUM-511, additional methods were added to the
"InProgressSubmission" interface which was causing the stock DSpace
"XmlWorkflowItem" implementation to fail. A UMD-customized version
of the "XmlWorkflowItem" class (in the "additions" module) which has
"stub" implementations of the methods has been restored.

The UMD-customized "InProgressSubmission" interface was also modified
in LIBDRUM-628, to support the Data Community workflow, which has not
yet been implemented in DSpace 7 (see LIBDRUM-682).

Since it is not clear whether the implementation of the Data Community
workflow will actually require the methods added in LIBDRUM-628, in this
issue the methods in the "nProgressSubmission" interface (and their
implementations in the "WorkspaceItem" and "XmlWorkflowItem" classes)
were commented out. They can be restored or removed as necessary in
LIBDRUM-682.

https://issues.umd.edu/browse/LIBDRUM-716