LIBDRUM-728. Using single "Describe" section for default submissions
Modified the default submission process to use a single "Describe"
section (instead of two "Describe" sections as in the "traditional"
submission process).

Instead of modifying the "traditional" submission process, created
a new "umd-submission" process in which the submission page has been
modified to use the "umdDescribe" step, which, in turn, uses the
"umdDescribe" form.

In the "umdDescribe" form:

* Moved "Type" field to be the first field, and made it required

* Added "External Link" field (based on field in
  “Minority Health and Health Equity Archive” submission form)

https://umd-dit.atlassian.net/browse/LIBDRUM-728