LIBDRUM-728. Using single "Describe" section for MHHEA submissions
Modified the "Minority Health and Health Equity Archive" submission
process to use a single "Describe" section (instead of two "Describe"
sections).

Renamed the "MHHEApageone" step/form to "MHHEADescribe", and copied all
the all questions from the "traditionalpagetwo" form into the
"MHHEADescribe" form.

Also, made the "Type" field required, and moved to be the first question
in the "MHHEADescribe" form, for consistency with the default submission
process form.

https://umd-dit.atlassian.net/browse/LIBDRUM-728