LIBDRUM-728. Share "Describe" section between default and MHHEA
The "Describe" section in the default (umd-submission) and MHHEA
submission forms has become identical. Based on discussions with Terry
Owen, it seems unlikely that this section will diverge in the future,
so combining them into a single configuration should simplify
maintenance and testing.

Removed the "MHHEADescribe" step definition and form, as they are no
longer needed.

Renamed the "MHHEA_Collection" submission process to "MHHEA-submission"
to be more in line with the default "umd-submission" process.

https://umd-dit.atlassian.net/browse/LIBDRUM-728