LIBDRUM-682. Added "Data Community" mapping to workflow
Added a "datacommunitycollectionmappingaction" Workflow action to
add a submission to the "UMD Data Community" collection, if the
submission is of type "Dataset" or "Software".

Updated an "data.community.handle" property in the
"config/local.cfg.EXAMPLE" file, to specify the handle of the
"UMD Data Community" community on the production server.

https://umd-dit.atlassian.net/browse/LIBDRUM-682