LIBDRUM-678. Added embargo restriction to BitstreamRest object
Added an "embargoRestriction" field to the  "BitstreamRest" object
passed between the back-end and frontend. The field is a simple string
that contains:

* "NONE" - no embargo (or embargo on item has been lifted)
* "FOREVER" - the embargo never ends
* A date (in yyyy-MM-dd) format, indicating the embargo lift date

Modified the "BitstreamConverter" to populate the "embargoRestriction"
field, based on any "ETD Embargo" resource policies on the bitstream.

Added "dspace/docs/DrumEmbargoAndAccessRestrictions.md" to record
information about the ETD Embargo functionality and other access
restrictions that affect the downloading of bitstreams.

https://umd-dit.atlassian.net/browse/LIBDRUM-678