LIBDRUM-738. Enable anonymous users to retrieve total downloads report
Created a custom copy of "UsageReportRestPermissionEvaluatorPlugin" and
modified the "hasDSpacePermission" to allow requests for the
"_TotalDownloads" usage report to skip the check of the
"usage-statistics.authorization.admin.usage" and instead determine
access based on whether the user has READ access to the bitstream.

This change is necessary to enable anonymous users to see the
"No. of downloads" for a bitstream on the "item" page, without
giving them access to the "Statistics" menu entry in the navigation bar.

https://umd-dit.atlassian.net/browse/LIBDRUM-738